[app] get app state from routes
